impl: implemented frame rendering system

+++ data.zig
@@ -1,4 +1,4 @@
 // File for program data structures
 
 pub const TermSize = struct { height: usize, width: usize };
-pub const Entity = struct { char: u8, x: u32, y: u32 };
+pub const Entity = struct { char: u8, x: usize, y: usize };
